This commit is a bit large - woops
I completely lost track of time and logged a few hours before I
felt like i was at a good place to commit. A lot of refactorings
on the background task. Figuring the best way to have the db session
available to the task and the routes.

Started on the frontend, got some lightweight css and some basic
ui/ux javascript going on. Next commit will integrate with neo through
bolt.